+++ chatbot/lib/middleware/wit.middleware.js
