+++ chatbot/lib/responses/slack.responses.js
@@ -1,0 +1,6 @@
+module.exports = {
+  bye : ['See you later!', 'Bye for now!'],
+  emoji : [':eyes:', ':heart:', ':slightly_smiling_face:', ':sleeping:'],
+  greeting : ['Hello!', 'How are you doing?'],
+  random : ['poke']
+};
